Create namespace for custom instrumentation tests
