+++ tests/Agent/IntegrationTests/IntegrationTests/CustomInstrumentation/CustomInstrumentationAsync.cs
@@ -9,7 +9,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.CustomInstrumentation
 {
     [NetFrameworkTest]
     public class CustomInstrumentationAsync : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
