+++ tests/Agent/IntegrationTests/IntegrationTests/CustomInstrumentation/OtherTransactionAsync.cs
@@ -9,7 +9,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.CustomInstrumentation
 {
     [NetFrameworkTest]
     public class OtherTransactionAsync : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
